Add new demo to draw a triangle using pbkit
This patch adds a new demo which simply draws a triangle using pbkit.
It is a highly simplified version of what is in the other pbkit demos.
It's relatively small and linear, but still covers enough to show how
shaders can be set up, attributes can be set, and geometry can be
rendered.

Tested on real Xbox and XQEMU.